Move RunRC-like functions to test/utils
