+++ test/e2e/cluster_size_autoscaling.go
@@ -31,6 +31,7 @@ import (
 	"k8s.io/kubernetes/pkg/fields"
 	"k8s.io/kubernetes/pkg/util/sets"
 	"k8s.io/kubernetes/test/e2e/framework"
+	testutils "k8s.io/kubernetes/test/utils"
 
 	"github.com/golang/glog"
 	. "github.com/onsi/ginkgo"
@@ -456,7 +457,7 @@ func doPut(url, content string) (string, error) {
 func CreateNodeSelectorPods(f *framework.Framework, id string, replicas int, nodeSelector map[string]string, expectRunning bool) {
 	By(fmt.Sprintf("Running RC which reserves host port and defines node selector"))
 
-	config := &framework.RCConfig{
+	config := &testutils.RCConfig{
 		Client:       f.Client,
 		Name:         "node-selector",
 		Namespace:    f.Namespace.Name,
@@ -474,7 +475,7 @@ func CreateNodeSelectorPods(f *framework.Framework, id string, replicas int, nod
 
 func CreateHostPortPods(f *framework.Framework, id string, replicas int, expectRunning bool) {
 	By(fmt.Sprintf("Running RC which reserves host port"))
-	config := &framework.RCConfig{
+	config := &testutils.RCConfig{
 		Client:    f.Client,
 		Name:      id,
 		Namespace: f.Namespace.Name,
@@ -492,7 +493,7 @@ func CreateHostPortPods(f *framework.Framework, id string, replicas int, expectR
 func ReserveCpu(f *framework.Framework, id string, replicas, millicores int) {
 	By(fmt.Sprintf("Running RC which reserves %v millicores", millicores))
 	request := int64(millicores / replicas)
-	config := &framework.RCConfig{
+	config := &testutils.RCConfig{
 		Client:     f.Client,
 		Name:       id,
 		Namespace:  f.Namespace.Name,
@@ -507,7 +508,7 @@ func ReserveCpu(f *framework.Framework, id string, replicas, millicores int) {
 func ReserveMemory(f *framework.Framework, id string, replicas, megabytes int, expectRunning bool) {
 	By(fmt.Sprintf("Running RC which reserves %v MB of memory", megabytes))
 	request := int64(1024 * 1024 * megabytes / replicas)
-	config := &framework.RCConfig{
+	config := &testutils.RCConfig{
 		Client:     f.Client,
 		Name:       id,
 		Namespace:  f.Namespace.Name,
