+++ test/e2e/load.go
@@ -37,6 +37,7 @@ import (
 	"k8s.io/kubernetes/pkg/util/intstr"
 	utilnet "k8s.io/kubernetes/pkg/util/net"
 	"k8s.io/kubernetes/test/e2e/framework"
+	testutils "k8s.io/kubernetes/test/utils"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -66,7 +67,7 @@ var _ = framework.KubeDescribe("Load capacity", func() {
 	var c *client.Client
 	var nodeCount int
 	var ns string
-	var configs []*framework.RCConfig
+	var configs []*testutils.RCConfig
 	var namespaces []*api.Namespace
 
 	// Gathers metrics before teardown
@@ -283,8 +284,8 @@ func computeRCCounts(total int) (int, int, int) {
 	return smallRCCount, mediumRCCount, bigRCCount
 }
 
-func generateRCConfigs(totalPods int, image string, command []string, nss []*api.Namespace) []*framework.RCConfig {
-	configs := make([]*framework.RCConfig, 0)
+func generateRCConfigs(totalPods int, image string, command []string, nss []*api.Namespace) []*testutils.RCConfig {
+	configs := make([]*testutils.RCConfig, 0)
 
 	smallRCCount, mediumRCCount, bigRCCount := computeRCCounts(totalPods)
 	configs = append(configs, generateRCConfigsForGroup(nss, smallRCGroupName, smallRCSize, smallRCCount, image, command)...)
@@ -305,10 +306,10 @@ func generateRCConfigs(totalPods int, image string, command []string, nss []*api
 }
 
 func generateRCConfigsForGroup(
-	nss []*api.Namespace, groupName string, size, count int, image string, command []string) []*framework.RCConfig {
-	configs := make([]*framework.RCConfig, 0, count)
+	nss []*api.Namespace, groupName string, size, count int, image string, command []string) []*testutils.RCConfig {
+	configs := make([]*testutils.RCConfig, 0, count)
 	for i := 1; i <= count; i++ {
-		config := &framework.RCConfig{
+		config := &testutils.RCConfig{
 			Client:     nil, // this will be overwritten later
 			Name:       groupName + "-" + strconv.Itoa(i),
 			Namespace:  nss[i%len(nss)].Name,
@@ -324,7 +325,7 @@ func generateRCConfigsForGroup(
 	return configs
 }
 
-func generateServicesForConfigs(configs []*framework.RCConfig) []*api.Service {
+func generateServicesForConfigs(configs []*testutils.RCConfig) []*api.Service {
 	services := make([]*api.Service, 0, len(configs))
 	for _, config := range configs {
 		serviceName := config.Name + "-svc"
@@ -351,7 +352,7 @@ func sleepUpTo(d time.Duration) {
 	time.Sleep(time.Duration(rand.Int63n(d.Nanoseconds())))
 }
 
-func createAllRC(configs []*framework.RCConfig, creatingTime time.Duration) {
+func createAllRC(configs []*testutils.RCConfig, creatingTime time.Duration) {
 	var wg sync.WaitGroup
 	wg.Add(len(configs))
 	for _, config := range configs {
@@ -360,7 +361,7 @@ func createAllRC(configs []*framework.RCConfig, creatingTime time.Duration) {
 	wg.Wait()
 }
 
-func createRC(wg *sync.WaitGroup, config *framework.RCConfig, creatingTime time.Duration) {
+func createRC(wg *sync.WaitGroup, config *testutils.RCConfig, creatingTime time.Duration) {
 	defer GinkgoRecover()
 	defer wg.Done()
 
@@ -368,7 +369,7 @@ func createRC(wg *sync.WaitGroup, config *framework.RCConfig, creatingTime time.
 	framework.ExpectNoError(framework.RunRC(*config), fmt.Sprintf("creating rc %s", config.Name))
 }
 
-func scaleAllRC(configs []*framework.RCConfig, scalingTime time.Duration) {
+func scaleAllRC(configs []*testutils.RCConfig, scalingTime time.Duration) {
 	var wg sync.WaitGroup
 	wg.Add(len(configs))
 	for _, config := range configs {
@@ -379,7 +380,7 @@ func scaleAllRC(configs []*framework.RCConfig, scalingTime time.Duration) {
 
 // Scales RC to a random size within [0.5*size, 1.5*size] and lists all the pods afterwards.
 // Scaling happens always based on original size, not the current size.
-func scaleRC(wg *sync.WaitGroup, config *framework.RCConfig, scalingTime time.Duration) {
+func scaleRC(wg *sync.WaitGroup, config *testutils.RCConfig, scalingTime time.Duration) {
 	defer GinkgoRecover()
 	defer wg.Done()
 
@@ -396,7 +397,7 @@ func scaleRC(wg *sync.WaitGroup, config *framework.RCConfig, scalingTime time.Du
 	framework.ExpectNoError(err, fmt.Sprintf("listing pods from rc %v", config.Name))
 }
 
-func deleteAllRC(configs []*framework.RCConfig, deletingTime time.Duration) {
+func deleteAllRC(configs []*testutils.RCConfig, deletingTime time.Duration) {
 	var wg sync.WaitGroup
 	wg.Add(len(configs))
 	for _, config := range configs {
@@ -405,7 +406,7 @@ func deleteAllRC(configs []*framework.RCConfig, deletingTime time.Duration) {
 	wg.Wait()
 }
 
-func deleteRC(wg *sync.WaitGroup, config *framework.RCConfig, deletingTime time.Duration) {
+func deleteRC(wg *sync.WaitGroup, config *testutils.RCConfig, deletingTime time.Duration) {
 	defer GinkgoRecover()
 	defer wg.Done()
